Update Question
Add question to sync or not

+++ src/portfolio/controllers/CheckConnection.java
@@ -3,6 +3,8 @@ package portfolio.controllers;
 import javafx.application.Platform;
 
 import java.io.File;
+import java.io.FileWriter;
+import java.io.IOException;
 import java.util.TimerTask;
 
 public class CheckConnection extends TimerTask {
@@ -23,16 +25,33 @@ public class CheckConnection extends TimerTask {
                             } else {
                                 SettingsController.getInstance().runCheckTimer = false;
                                 SettingsController.getInstance().errorBouncer = 0;
-                                File file = new File(System.getProperty("user.dir") + "/PortfolioData/"+"update.portfolio");
-                                if(file.exists())file.delete();
+                                File file = new File(System.getProperty("user.dir") + "/PortfolioData/" + "update.portfolio");
+                                if (file.exists()) file.delete();
                             }
                         } else {
-                            SettingsController.getInstance().runCheckTimer = false;
-                            SettingsController.getInstance().errorBouncer = 0;
-                            this.mainViewController.btnUpdateDatabasePressed();
-                            this.mainViewController.plotUpdate(this.mainViewController.mainView.tabPane.getSelectionModel().getSelectedItem().getText());
-                            File file = new File(System.getProperty("user.dir") + "/PortfolioData/"+"update.portfolio");
-                            if(file.exists())file.delete();
+
+                            if (this.mainViewController.settingsController.selectedLaunchSync) {
+                                SettingsController.getInstance().runCheckTimer = false;
+                                SettingsController.getInstance().errorBouncer = 0;
+                                this.mainViewController.btnUpdateDatabasePressed();
+                                this.mainViewController.plotUpdate(this.mainViewController.mainView.tabPane.getSelectionModel().getSelectedItem().getText());
+                                File file = new File(System.getProperty("user.dir") + "/PortfolioData/" + "update.portfolio");
+                                if (file.exists()) file.delete();
+                            }else{
+                                int currentBlockCount = Integer.parseInt(this.mainViewController.transactionController.getBlockCountRpc());
+                                int maxBlockCount = Integer.parseInt(this.mainViewController.transactionController.getBlockCount());
+                                if(SettingsController.getInstance().getPlatform().equals("mac")){
+                                    try {
+                                        FileWriter myWriter = new FileWriter(System.getProperty("user.dir") + "/PortfolioData/"+"update.portfolio");
+                                        myWriter.write(SettingsController.getInstance().translationList.getValue().get("SyncData").toString() +"("+currentBlockCount+"/"+maxBlockCount+")");
+                                        myWriter.close();
+                                    } catch (IOException e) {
+                                        SettingsController.getInstance().logger.warning("Could not write to update.portfolio."); }
+                                }else{
+                                    TransactionController.getInstance().jl.setText(SettingsController.getInstance().translationList.getValue().get("SyncData").toString() +"("+currentBlockCount+"/"+maxBlockCount+")");
+                                }
+                                this.mainViewController.settingsController.selectedLaunchSync = currentBlockCount>=maxBlockCount;
+                            }
                         }
                     }
                 }
